requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-telegram-bot==22.0
psycopg[binary]==3.2.6
pydantic==2.11.3
python-dotenv==1.1.0
httpx==0.28.1
